+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -20,6 +20,11 @@ import java.util.UUID;
 import org.apache.catalina.manager.Session;
 import org.apache.catalina.manager.SessionManager;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpCookie;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.StatusCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
